+++ economy/views.py
@@ -4,6 +4,8 @@ import stripe
 from django.conf import settings
 from django.db import transaction
 from graphene_django_cud.util import disambiguate_id, disambiguate_ids
+
+from common.decorators import view_feature_flag_required
 from economy.models import (
     SociProduct,
     Deposit,
@@ -77,7 +79,6 @@ def stripe_webhook(request):
         deposit = Deposit.objects.get(stripe_payment_id=intent_id)
         if deposit.approved:
             # Already approved. Do nothing
-
             return JsonResponse(data={"success": True})
 
         with transaction.atomic():
@@ -89,6 +90,33 @@ def stripe_webhook(request):
             deposit.account.add_funds(deposit.resolved_amount)
             if deposit.account.user.notify_on_deposit:
                 send_deposit_approved_email(deposit)
+
+    elif event["type"] == "charge.refunded":
+        event_object = event["data"]["object"]
+        payment_intent_id = event_object["payment_intent"]
+
+        deposit = Deposit.objects.get(stripe_payment_id=payment_intent_id)
+
+        if not deposit.approved:
+            # Already invalidated. Do nothing
+            return JsonResponse(data={"success": True})
+
+        amount_captured = event_object["amount_captured"]
+        amount_refunded = event_object["amount_refunded"]
+        amount = event_object["amount"]
+
+        calculated_fee = deposit.amount - deposit.resolved_amount
+
+        with transaction.atomic():
+            from economy.utils import send_deposit_refunded_email
+
+            deposit.approved = False
+            deposit.account.remove_funds(deposit.resolved_amount)
+            deposit.save()
+            if deposit.account.user.notify_on_deposit:
+                send_deposit_refunded_email(deposit)
+                # Could be confusing user flow if we don't delete the deposit
+                deposit.delete()
     else:
         raise Exception(f"Unhandled event type {event['type']}")